+++ src/client/cypress/e2e/detailPage/sideNavigation.cy.js
@@ -4,11 +4,12 @@ import {
   BoreholeTab,
   isActiveMenuItem,
   isActiveTab,
-  isInactiveBoreholeTab,
-  isInactiveTab,
+  isInactiveEmptyTab,
+  isInactiveTabWithContent,
   isMenuItemWithContent,
   isMenuItemWithoutContent,
   navigateInSidebar,
+  navigateToTabWithTitle,
   SidebarMenuItem,
   StratigraphyTab,
 } from "../helpers/navigationHelpers.js";
@@ -18,6 +19,13 @@ import {
   selectInputFile,
   startBoreholeEditing,
 } from "../helpers/testHelpers";
+
+function addHierarchicalLayer(stratigraphyType) {
+  cy.get('[data-cy="add-layer-button"]').click({ force: true });
+  cy.wait(`@${stratigraphyType}_POST`);
+  cy.get('[data-cy="layer-card"] [data-testid="EditIcon"]').click();
+  cy.get('[data-cy="layer-card"] :nth-child(2) > .MuiAutocomplete-root').click();
+}
 
 describe("Test for the detail page side navigation.", () => {
   it("tests if navigation points are greyed out if there is no content", () => {
@@ -41,8 +49,8 @@ describe("Test for the detail page side navigation.", () => {
     // Check borehole content tabs
     navigateInSidebar(SidebarMenuItem.borehole);
     cy.wait("@codelist_GET");
-    isInactiveBoreholeTab(BoreholeTab.sections, false);
-    isInactiveBoreholeTab(BoreholeTab.geometry, false);
+    isInactiveEmptyTab(BoreholeTab.sections);
+    isInactiveEmptyTab(BoreholeTab.geometry);
 
     // Check empty Stratigraphy
     navigateInSidebar(SidebarMenuItem.stratigraphy);
@@ -62,7 +70,6 @@ describe("Test for the detail page side navigation.", () => {
     addItem("addEmptyStratigraphy");
     setInput("name", "AUTODESPERADO");
     saveWithSaveBar();
-    cy.wait(["@lithology_by_stratigraphyId_GET", "@lithological_description", "@facies_description"]);
     isActiveMenuItem(SidebarMenuItem.stratigraphy, true);
 
     // Add completion
@@ -92,9 +99,29 @@ describe("Test for the detail page side navigation.", () => {
     isActiveMenuItem(SidebarMenuItem.completion, true);
     isMenuItemWithContent(SidebarMenuItem.stratigraphy);
     navigateInSidebar(SidebarMenuItem.stratigraphy);
-    isActiveTab(StratigraphyTab.lithology + "-tab");
-    isInactiveTab(StratigraphyTab.chronostratigraphy + "-tab");
-    isInactiveTab(StratigraphyTab.lithostratigraphy + "-tab");
+    isActiveTab(StratigraphyTab.lithology);
+    isInactiveEmptyTab(StratigraphyTab.chronostratigraphy);
+    isInactiveEmptyTab(StratigraphyTab.lithostratigraphy);
+
+    // add chronostratigraphy and lithostratigraphy to check if they are enabled when content is present
+    navigateToTabWithTitle("Chronostratigraphy");
+    addHierarchicalLayer("chronostratigraphy");
+    isInactiveEmptyTab(StratigraphyTab.lithology);
+    isActiveTab(StratigraphyTab.chronostratigraphy);
+    isInactiveEmptyTab(StratigraphyTab.lithostratigraphy);
+
+    navigateToTabWithTitle("Lithology");
+    isInactiveTabWithContent(StratigraphyTab.chronostratigraphy);
+    isInactiveEmptyTab(StratigraphyTab.lithostratigraphy);
+
+    navigateToTabWithTitle("Lithostratigraphy");
+    addHierarchicalLayer("lithostratigraphy");
+    isInactiveEmptyTab(StratigraphyTab.lithology);
+    isInactiveTabWithContent(StratigraphyTab.chronostratigraphy);
+    isActiveTab(StratigraphyTab.lithostratigraphy);
+    navigateToTabWithTitle("Lithology");
+    isInactiveTabWithContent(StratigraphyTab.chronostratigraphy);
+    isInactiveTabWithContent(StratigraphyTab.lithostratigraphy);
 
     navigateInSidebar(SidebarMenuItem.hydrogeology);
     navigateInSidebar(SidebarMenuItem.waterIngress);